feat(snippets): add FormTextFields and FormMenuFields components for enhanced form handling
