+++ src/types/fieldGropProps.ts
@@ -1,0 +1,6 @@
+import { EditInfo } from '@/types/snippets'
+
+export interface FieldGroupProps {
+  editInfo: EditInfo;
+  onChange: (updates: { [key: string]: string | string[] | boolean }) => void;
+}
